split client/server proxy config
