+++ controllers/prefix/prefix_reconciler.go
@@ -38,7 +38,7 @@ import (
 	"github.com/nokia/k8s-ipam/internal/meta"
 	"github.com/nokia/k8s-ipam/internal/resource"
 	"github.com/nokia/k8s-ipam/internal/shared"
-	"github.com/nokia/k8s-ipam/pkg/ipamproxy"
+	"github.com/nokia/k8s-ipam/pkg/clientipamproxy"
 	"github.com/pkg/errors"
 )
 
@@ -88,8 +88,7 @@ func Setup(mgr ctrl.Manager, options *shared.Options) (schema.GroupVersionKind, 
 type reconciler struct {
 	client.Client
 	Scheme          *runtime.Scheme
-	IpamClientProxy ipamproxy.IpamClientProxy
-	//Ipam         ipam.Ipam
+	IpamClientProxy clientipamproxy.Proxy
 	pollInterval time.Duration
 	finalizer    *resource.APIFinalizer
