+++ internal/shared/options.go
@@ -22,7 +22,7 @@ import (
 	"github.com/henderiw-k8s-lcnc/discovery/registrator"
 	"github.com/nokia/k8s-ipam/internal/injectors"
 	"github.com/nokia/k8s-ipam/internal/ipam"
-	"github.com/nokia/k8s-ipam/pkg/ipamproxy"
+	"github.com/nokia/k8s-ipam/pkg/clientipamproxy"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller"
 )
@@ -30,7 +30,7 @@ import (
 type Options struct {
 	PorchClient     client.Client
 	Registrator     registrator.Registrator
-	IpamClientProxy ipamproxy.IpamClientProxy
+	IpamClientProxy clientipamproxy.Proxy
 	Poll            time.Duration
 	Copts           controller.Options
 	Ipam            ipam.Ipam
